requirements.txt
fastapi==0.103.1
numpy==1.25.2
python-dotenv==1.0.0
pytube==15.0.0
Requests==2.31.0
streamlit==1.26.0
streamlit_card==0.0.61
